Update to 3.0.5 - Updated dependencies - Fix cleaning backups deleting all backups [#68]

+++ core/components/backupmodx/src/BackupMODX.php
@@ -40,7 +40,7 @@ class BackupMODX
      * The version
      * @var string $version
      */
-    public $version = '3.0.4-pl';
+    public $version = '3.0.5-pl';
 
     /**
      * The class options
@@ -93,7 +93,7 @@ class BackupMODX
             'mysqldumpAlias' => $this->getOption('mysqldumpAlias', $options, 'mysqldump'),
             'debug' => (bool)$this->getOption('debug', $options, false),
             'logTarget' => json_decode($this->getOption('logTarget', $options, '{"target":"FILE","options":{"filename":"' . $this->namespace . '.log"}}'), true),
-            'timelimit' => (int) $this->getOption('timelimit', $options, 120)
+            'timelimit' => (int)$this->getOption('timelimit', $options, 120)
         ));
 
         $lexicon = $this->modx->getService('lexicon', 'modLexicon');
@@ -200,6 +200,13 @@ class BackupMODX
                     'note' => $note,
                     'files' => $files,
                 );
+            } else {
+                $backups[] = array(
+                    'name' => $folder,
+                    'path' => $this->getOption('targetPath'),
+                    'date' => filemtime($this->getOption('targetPath') . $folder),
+                    'date_format' => date($this->modx->getOption('manager_date_format') . ' ' . $this->modx->getOption('manager_time_format'), filemtime($this->getOption('targetPath') . $folder)),
+                );
             }
         }
 
@@ -248,23 +255,27 @@ class BackupMODX
 
         // Remove old files
         foreach (array_slice($backups, $maxFiles) as $backup) {
-            $file = $backup['path'] . $backup['filename'] . '/' . $backup['files']['files'];
-            if (file_exists($file)) {
-                unlink($file);
+            if (isset($backup['files']) && isset($backup['files']['files'])) {
+                $file = $backup['path'] . $backup['name'] . '/' . $backup['files']['files'];
+                if (file_exists($file)) {
+                    unlink($file);
+                }
             }
         }
 
         // Remove old databases
         foreach (array_slice($backups, $maxDatabase) as $backup) {
-            $file = $backup['path'] . $backup['filename'] . '/' . $backup['files']['database'];
-            if (file_exists($file)) {
-                unlink($file);
+            if (isset($backup['files']) && isset($backup['files']['database'])) {
+                $file = $backup['path'] . $backup['name'] . '/' . $backup['files']['database'];
+                if (file_exists($file)) {
+                    unlink($file);
+                }
             }
         }
 
         // Remove old backup folders
         foreach (array_slice($backups, $max) as $backup) {
-            $this->removeBackup($backup['filename']);
+            $this->removeBackup($backup['name']);
         }
     }
 }
